requirements.txt
opencv-python
scikit-image
Pillow
imageio
tensorflow
lycon
jpeg4py
torch
torchvision